Refactor button into smaller components
