+++ js/components/HomeScene/Button/Button.module.scss
@@ -1,30 +1,3 @@
-.arrows-wrapper {
-    position: absolute;
-    bottom: var(--margin);
-    left: var(--margin);
-}
-
-.back-wrapper {
-    position: absolute;
-    bottom: var(--margin);
-    left: var(--margin);
-    opacity: 0;
-}
-
-.button-wrapper {
-    position: relative;
-    display: flex;
-    gap: 4px;
-    flex-flow: row nowrap;
-    align-items: end;
-}
-
-.middle-buttons {
-    display: flex;
-    flex-flow: column nowrap;
-    gap: 4px;
-}
-
 button.button {
     border: 1px solid;
     width: 50px;
