+++ js/components/HomeScene/Button/constants.js
@@ -1,0 +1,7 @@
+export const BUTTON_THEMES = {
+    white: "white",
+    black: "black",
+    yellow: "yellow",
+    orange: "orange",
+    violet: "violet",
+};
