+++ js/components/HomeScene/Button/index.jsx
@@ -1,0 +1,3 @@
+export { default } from "./Button";
+// eslint-disable-next-line
+export * from "./constants";
