+++ js/components/HomeScene/Nav/Nav.jsx
@@ -1,163 +1,11 @@
 import styles from "./Nav.module.scss";
 
-import { useLayoutEffect, useMemo, useRef, useState } from "react";
-import { useHotkeys } from "react-hotkeys-hook";
-import cx from "classnames";
+import { useLayoutEffect, useMemo, useRef } from "react";
 
 import gsap from "@/js/lib/gsap";
 import { useStore } from "@/js/lib/store";
 import Actions from "./Actions";
-
-const LeftArrow = () => {
-    return (
-        <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="16"
-            height="12"
-            viewBox="0 0 16 12"
-            fill="none"
-        >
-            <path
-                d="M16 5.99995L1.61749 5.99995M1.61749 5.99995L6.68079 11.0633M1.61749 5.99995L6.68079 0.936645"
-                stroke="currentColor"
-            />
-        </svg>
-    );
-};
-
-const UpArrow = () => {
-    return (
-        <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="12"
-            height="16"
-            viewBox="0 0 12 16"
-            fill="none"
-        >
-            <path
-                d="M5.99919 16L5.99919 1.61749M5.99919 1.61749L0.935891 6.68079M5.99919 1.61749L11.0625 6.68079"
-                stroke="currentColor"
-            />
-        </svg>
-    );
-};
-
-const DownArrow = () => {
-    return (
-        <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="12"
-            height="16"
-            viewBox="0 0 12 16"
-            fill="none"
-            transform="rotate(180 0 0)"
-        >
-            <path
-                d="M5.99919 16L5.99919 1.61749M5.99919 1.61749L0.935891 6.68079M5.99919 1.61749L11.0625 6.68079"
-                stroke="currentColor"
-            />
-        </svg>
-    );
-};
-
-const RightArrow = () => {
-    return (
-        <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="16"
-            height="12"
-            viewBox="0 0 16 12"
-            fill="none"
-        >
-            <path
-                d="M6.03793e-08 5.99995L14.3825 5.99995M14.3825 5.99995L9.31921 11.0633M14.3825 5.99995L9.31921 0.936645"
-                stroke="currentColor"
-            />
-        </svg>
-    );
-};
-
-function randomNumber(min, max) {
-    return Math.random() * (max - min) + min;
-}
-
-const ARROWS = {
-    left: LeftArrow,
-    up: UpArrow,
-    right: RightArrow,
-    down: DownArrow,
-};
-
-const COLORS = ["yellow", "orange", "violet"];
-const ACTIVE_COLOR = "white";
-
-const randomColorIndex = () => Math.floor(randomNumber(0, COLORS.length));
-
-const Button = ({
-    direction = "left",
-    onClick = () => {},
-    startingColorIndex = randomColorIndex(),
-    hotkey = null,
-}) => {
-    const [isActive, setIsActive] = useState(false);
-    const [colorIndex, setColorIndex] = useState(null);
-    const [lastColorIndex, setLastColorIndex] = useState(startingColorIndex);
-    const Component = ARROWS[direction];
-    const _hotkey = hotkey ?? direction;
-
-    const colorClassname = useMemo(() => {
-        if (isActive) {
-            return styles[`button--${ACTIVE_COLOR}`];
-        }
-        if (typeof colorIndex === "number") {
-            return styles[`button--${COLORS[colorIndex]}`];
-        }
-        return null;
-    }, [colorIndex, isActive]);
-
-    const handleMouseEnter = () => {
-        const newColorIndex =
-            lastColorIndex + 1 >= COLORS.length ? 0 : lastColorIndex + 1;
-        setColorIndex(newColorIndex);
-    };
-
-    const handleMouseLeave = () => {
-        setLastColorIndex(colorIndex);
-        setColorIndex(null);
-    };
-
-    const handleMousedown = () => {
-        setIsActive(true);
-    };
-
-    const handleMouseup = () => {
-        setTimeout(() => {
-            setIsActive(false);
-        }, 100);
-    };
-
-    useHotkeys(_hotkey, () => {
-        setIsActive(true);
-        setTimeout(() => {
-            setIsActive(false);
-        }, 100);
-    });
-
-    return (
-        <button
-            className={cx(styles.button, colorClassname)}
-            onClick={() => {
-                onClick(direction);
-            }}
-            onMouseEnter={handleMouseEnter}
-            onMouseLeave={handleMouseLeave}
-            onMouseDown={handleMousedown}
-            onMouseUp={handleMouseup}
-        >
-            <Component />
-        </button>
-    );
-};
+import NavButton from "./NavButton";
 
 const NavArrows = ({ visible }) => {
     const el = useRef(null);
@@ -190,12 +38,12 @@ const NavArrows = ({ visible }) => {
     return (
         <div className={styles.arrowsWrapper} ref={el}>
             <div className={styles.buttonWrapper}>
-                <Button direction="left" onClick={orbit} />
+                <NavButton direction="left" onClick={orbit} />
                 <div className={styles.middleButtons}>
-                    <Button direction="up" onClick={orbit} />
-                    <Button direction="down" onClick={orbit} />
+                    <NavButton direction="up" onClick={orbit} />
+                    <NavButton direction="down" onClick={orbit} />
                 </div>
-                <Button direction="right" onClick={orbit} />
+                <NavButton direction="right" onClick={orbit} />
             </div>
         </div>
     );
@@ -231,7 +79,7 @@ const NavBack = ({ visible }) => {
     }, [visible]);
     return (
         <div className={styles.backWrapper} ref={el}>
-            <Button
+            <NavButton
                 direction="left"
                 onClick={() => {
                     resetCurrentBoxUuid();
